+++ frontend/src/components/Login.js
@@ -3,6 +3,7 @@ import axios from 'axios';
 import jwt_decode from 'jwt-decode';
 import setAuthToken from '../utils/setAuthToken';
 import { Redirect } from 'react-router-dom';
+import { render } from '@testing-library/react';
 const REACT_APP_SERVER_URL = process.env.REACT_APP_SERVER_URL;
 console.log(REACT_APP_SERVER_URL)
